pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "langgraph-rag-guardrails"
version = "0.1.0"
description = "Advanced RAG reference implementation with LangGraph, hybrid retrieval, and guardrails"
requires-python = ">=3.11,<3.14"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    # Pydantic
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # LangChain / LangGraph
    "langgraph>=0.1.19",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-ollama>=0.1.0",
    "langchain-openai>=0.1.8",
    "langchain-community>=0.2.0",
    # AI / ML
    "sentence-transformers>=3.0.0",
    "torch>=2.2.0",
    # Vector DB
    "qdrant-client>=1.9.0",
    # Cache
    "redis[asyncio]>=5.0.0",
    # Sparse retrieval
    "rank-bm25>=0.2.2",
    # Rate limiting
    "slowapi>=0.1.9",
    # YAML
    "PyYAML>=6.0.1",
    # Markdown parsing
    "python-frontmatter>=1.1.0",
    # HTTP client (for tests)
    "httpx>=0.27.0",
    # Numpy
    "numpy>=1.26.0",
    # Observability (optional)
    "langfuse>=2.36.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "ingest*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
